Add blkin support for librbd aio_write
Add a rbd_aio_write_traced operation to the librbd API. This allows a client (e.g. fio rbd engine) to pass a blkin trace to librbd.

The blkin C headers were added to src/common/zipkin_c.h to provide the same functionality as the src/common/zipkin_trace.h for C++.

A trace_event function was added to ImageCtx, as well as a trace endpoint and a trace object, to record events using the trace information passed from the client.