Pass trace information for aio_writes within librbd.
A rbd_aio_write_traced function was created in the RBD API to be used from fio rbd engine.

A  blkin_trace_infomation* field was added to AioImageRequest, which passes the trace information to the cache requests. A similar field was added to the Bufferhead to be able to pass the trace information when an AioObjectRequest is created.

Signed-off-by: Victor Araujo <ve.ar91@gmail.com>